First run converting to ES2015. Working on private properties.
