+++ package.json
@@ -14,10 +14,18 @@
     "url": "https://github.com/Trakkasure/DataManager.git"
   },
   "devDependencies": {
+    "babel": "^6.5.2",
+    "babel-cli": "^6.10.1",
+    "babel-preset-es2015": "^6.9.0",
+    "babel-preset-stage-1": "^6.5.0",
     "mocha": ">=1.7.4",
-    "uglify-js": ">=1.2.5"
+    "private-members-babel-plugin": "git+https://github.com/ArikMaor/private-members-babel-plugin.git",
+    "uglify-js": ">=1.2.5",
+    "webpack": "^1.13.1",
+    "webpack-dev-server": "^1.14.1"
   },
   "dependencies": {
+    "crossfilter": "^1.3.12",
     "symbol-observable": "^1.0.1"
   }
 }
